flask-food-starter/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.1.1
pytest==8.4.2
